Refactor predicate classes
