+++ src/main/java/seedu/address/logic/commands/HidePatientsCommand.java
@@ -9,7 +9,7 @@ import java.util.function.Predicate;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.model.Model;
-import seedu.address.model.person.HiddenPredicateSingleton;
+import seedu.address.model.person.predicates.HiddenPredicateSingleton;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.predicates.AppointmentOfFilteredPersonsPredicate;
