+++ src/main/java/seedu/address/model/person/predicates/HiddenPredicateSingleton.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.person.predicates;
 
 
 import java.util.ArrayList;
@@ -6,6 +6,8 @@ import java.util.List;
 import java.util.function.Predicate;
 
 import seedu.address.model.Model;
+import seedu.address.model.person.Appointment;
+import seedu.address.model.person.Person;
 
 
 /**
