+++ src/main/java/seedu/address/model/person/predicates/NameContainsKeywordsPredicate.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.function.Predicate;
 
 import seedu.address.commons.util.StringUtil;
-import seedu.address.model.person.HiddenPredicateSingleton;
 import seedu.address.model.person.Person;
 
 /**
