+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -44,9 +44,9 @@ import seedu.address.logic.commands.MarkCommand;
 import seedu.address.logic.commands.UnmarkCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.person.Appointment;
-import seedu.address.model.person.HideAppointmentPredicate;
+import seedu.address.model.person.predicates.HideAppointmentPredicate;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.TagContainsKeywordsPredicate;
+import seedu.address.model.person.predicates.TagContainsKeywordsPredicate;
 import seedu.address.model.person.predicates.CombinedAppointmentPredicate;
 import seedu.address.model.person.predicates.CombinedPersonPredicate;
 import seedu.address.testutil.AppointmentUtil;
